requirements.txt
python-dotenv==1.0.0
selenium==4.18.1
telethon==1.34.0 